refactor(architecture): sink the configuration reader into the kernel (MOD-TODO wave)
src/dashboard/configuration.ts -> src/configuration.ts (MOD-SHARED-KERNEL):
the 13-line host configuration reader sat in the composition top, so the
todo service imported upwards and closed the DASHBOARD-SHELL<->TODO
2-cycle. After the move the todo module's remaining shell edges are
type-only, the 2-cycle dissolves, the module SCC cluster shrinks 11 -> 10,
and the todo entrypoints narrow to the consumed surface (service,
todoPanelCapability, types). Anchors in the webview checks, the dashboard
boundary contract test, and the performance-architecture baseline move with
the file. No behavior change.
