+++ .ci/architecture-debt-baseline.json
@@ -13,8 +13,7 @@
         "2:MOD-AI-SESSION-PROVIDER->MOD-WORKSPACE-IDENTITY",
         "2:MOD-DASHBOARD-SHELL->MOD-OPEN-WORKSPACE",
         "2:MOD-DASHBOARD-SHELL->MOD-SKILL-MANAGEMENT",
-        "2:MOD-DASHBOARD-SHELL->MOD-TODO",
-        "scc:MOD-AI-SESSION-ATTENTION|MOD-AI-SESSION-CONTROL|MOD-AI-SESSION-CONVERSATION|MOD-AI-SESSION-PROVIDER|MOD-AI-SESSION-RUNTIME|MOD-DASHBOARD-SHELL|MOD-OPEN-WORKSPACE|MOD-PROJECT-CATALOG|MOD-SKILL-MANAGEMENT|MOD-TODO|MOD-WORKSPACE-IDENTITY"
+        "scc:MOD-AI-SESSION-ATTENTION|MOD-AI-SESSION-CONTROL|MOD-AI-SESSION-CONVERSATION|MOD-AI-SESSION-PROVIDER|MOD-AI-SESSION-RUNTIME|MOD-DASHBOARD-SHELL|MOD-OPEN-WORKSPACE|MOD-PROJECT-CATALOG|MOD-SKILL-MANAGEMENT|MOD-WORKSPACE-IDENTITY"
       ]
     }
   }
